documentation and some optimizations added

+++ app/src/main/java/productions/darthplagueis/pit/view/PitPoint.java
@@ -3,34 +3,39 @@ package productions.darthplagueis.pit.view;
 import android.content.Context;
 import android.view.View;
 
+/**
+ * Class that is used create the point objects that will be manipulated
+ * by PitView. The X and Y field variables hold values that correspond
+ * to the point object's position in PitView.
+ */
 public class PitPoint extends View {
 
-    private float xPosition;
-    private float yPosition;
+    private int xPosition;
+    private int yPosition;
 
     public PitPoint(Context context) {
         super(context);
     }
 
-    public PitPoint(Context context, float xPosition, float yPosition) {
+    public PitPoint(Context context, int xPosition, int yPosition) {
         super(context);
         this.xPosition = xPosition;
         this.yPosition = yPosition;
     }
 
-    public float getXPosition() {
+    public int getXPosition() {
         return xPosition;
     }
 
-    public void setXPosition(float xPosition) {
+    public void setXPosition(int xPosition) {
         this.xPosition = xPosition;
     }
 
-    public float getYPosition() {
+    public int getYPosition() {
         return yPosition;
     }
 
-    public void setYPosition(float yPosition) {
+    public void setYPosition(int yPosition) {
         this.yPosition = yPosition;
     }
 }
